+++ tests/rx/cv_short.cpp
@@ -5,7 +5,11 @@ TEST_F(RxTest, cv_short_cv23) {
   auto packet{
     dcc::make_cv_access_short_write_packet(_addrs.primary, 0b0010u, cv23)};
 
-  EXPECT_CALL(_mock, writeCv(23u - 1u, cv23));
+  EXPECT_CALL(_mock,
+              writeCv(Matcher<uint32_t>(23u - 1u),
+                      Matcher<uint8_t>(cv23),
+                      Matcher<std::function<void(uint8_t)>>(_)))
+    .WillOnce(InvokeArgument<2uz>(cv23));
   if constexpr (DCC_STANDARD_COMPLIANCE) ReceiveAndExecute(packet);
   else ReceiveAndExecuteTwice(packet);
 }
@@ -16,7 +20,15 @@ TEST_F(RxTest, cv_short_cv31_32) {
   auto packet{dcc::make_cv_access_short_write_packet(
     _addrs.primary, 0b0101u, cv31, cv32)};
 
-  EXPECT_CALL(_mock, writeCv(31u - 1u, cv31));
-  EXPECT_CALL(_mock, writeCv(32u - 1u, cv32));
+  EXPECT_CALL(_mock,
+              writeCv(Matcher<uint32_t>(31u - 1u),
+                      Matcher<uint8_t>(cv31),
+                      Matcher<std::function<void(uint8_t)>>(_)))
+    .WillOnce(InvokeArgument<2uz>(cv31));
+  EXPECT_CALL(_mock,
+              writeCv(Matcher<uint32_t>(32u - 1u),
+                      Matcher<uint8_t>(cv32),
+                      Matcher<std::function<void(uint8_t)>>(_)))
+    .WillOnce(InvokeArgument<2uz>(cv32));
   ReceiveAndExecuteTwice(packet);
 }